feat: Migrating ledger allocation tool from personal repository.
This is a recent rewrite to condense things to the bare minimum. The old
version of this code had collected a ton of unused cruft from the ruins
of previous experiments. Right now this also has the least amount of
personal information baked into it than it ever has, haha.