feat: Experimental rebalancing tool.
I have some ideas about how to make this more useful than what's currently being
implemented here. Basically, this allows the user to define rules to apply to a
ledger file and the script will determine which transactions the user needs to
make and output those as uncleared ledger blocks.